+++ ratatui-core/src/buffer.rs
