+++ ratatui-core/src/buffer/assert.rs
