+++ ratatui-core/src/buffer/buffer.rs
@@ -23,7 +23,7 @@ use crate::{
 /// # Examples:
 ///
 /// ```
-/// use ratatui::{
+/// use ratatui_core::{
 ///     buffer::{Buffer, Cell},
 ///     layout::{Position, Rect},
 ///     style::{Color, Style},
@@ -168,7 +168,7 @@ impl Buffer {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     buffer::{Buffer, Cell},
     ///     layout::{Position, Rect},
     /// };
@@ -199,7 +199,7 @@ impl Buffer {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     buffer::{Buffer, Cell},
     ///     layout::{Position, Rect},
     ///     style::{Color, Style},
@@ -227,7 +227,7 @@ impl Buffer {
     /// # Examples
     ///
     /// ```
-    /// use ratatui::{buffer::Buffer, layout::Rect};
+    /// use ratatui_core::{buffer::Buffer, layout::Rect};
     ///
     /// let buffer = Buffer::empty(Rect::new(200, 100, 10, 10));
     /// // Global coordinates to the top corner of this buffer's area
@@ -239,7 +239,7 @@ impl Buffer {
     /// Panics when given an coordinate that is outside of this Buffer's area.
     ///
     /// ```should_panic
-    /// use ratatui::{buffer::Buffer, layout::Rect};
+    /// use ratatui_core::{buffer::Buffer, layout::Rect};
     ///
     /// let buffer = Buffer::empty(Rect::new(200, 100, 10, 10));
     /// // Top coordinate is outside of the buffer in global coordinate space, as the Buffer's area
@@ -282,7 +282,7 @@ impl Buffer {
     /// # Examples
     ///
     /// ```
-    /// use ratatui::{buffer::Buffer, layout::Rect};
+    /// use ratatui_core::{buffer::Buffer, layout::Rect};
     ///
     /// let rect = Rect::new(200, 100, 10, 10);
     /// let buffer = Buffer::empty(rect);
@@ -295,7 +295,7 @@ impl Buffer {
     /// Panics when given an index that is outside the Buffer's content.
     ///
     /// ```should_panic
-    /// use ratatui::{buffer::Buffer, layout::Rect};
+    /// use ratatui_core::{buffer::Buffer, layout::Rect};
     ///
     /// let rect = Rect::new(0, 0, 10, 10); // 100 cells in total
     /// let buffer = Buffer::empty(rect);
@@ -526,7 +526,7 @@ impl<P: Into<Position>> Index<P> for Buffer {
     /// # Examples
     ///
     /// ```
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     buffer::{Buffer, Cell},
     ///     layout::{Position, Rect},
     /// };
@@ -556,7 +556,7 @@ impl<P: Into<Position>> IndexMut<P> for Buffer {
     /// # Examples
     ///
     /// ```
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     buffer::{Buffer, Cell},
     ///     layout::{Position, Rect},
     /// };
