+++ ratatui-core/src/buffer/cell.rs
