+++ ratatui-core/src/layout/rect/iter.rs
