+++ ratatui-core/src/style/palette_conversion.rs
@@ -15,7 +15,7 @@ use crate::style::Color;
 ///
 /// ```
 /// use palette::Srgb;
-/// use ratatui::style::Color;
+/// use ratatui_core::style::Color;
 ///
 /// let color = Color::from(Srgb::new(1.0f32, 0.0, 0.0));
 /// assert_eq!(color, Color::Rgb(255, 0, 0));
@@ -36,7 +36,7 @@ impl<T: IntoStimulus<u8>> From<Srgb<T>> for Color {
 ///
 /// ```
 /// use palette::LinSrgb;
-/// use ratatui::style::Color;
+/// use ratatui_core::style::Color;
 ///
 /// let color = Color::from(LinSrgb::new(1.0f32, 0.0, 0.0));
 /// assert_eq!(color, Color::Rgb(255, 0, 0));
