+++ ratatui-core/src/text/line.rs
@@ -75,7 +75,7 @@ use crate::{
 /// [`Style`].
 ///
 /// ```rust
-/// use ratatui::{
+/// use ratatui_core::{
 ///     style::{Color, Modifier, Style, Stylize},
 ///     text::{Line, Span},
 /// };
@@ -102,7 +102,7 @@ use crate::{
 /// methods of the [`Stylize`] trait.
 ///
 /// ```rust
-/// use ratatui::{
+/// use ratatui_core::{
 ///     style::{Color, Modifier, Style, Stylize},
 ///     text::Line,
 /// };
@@ -121,7 +121,7 @@ use crate::{
 /// ignored and the line is truncated.
 ///
 /// ```rust
-/// use ratatui::{layout::Alignment, text::Line};
+/// use ratatui_core::{layout::Alignment, text::Line};
 ///
 /// let line = Line::from("Hello world!").alignment(Alignment::Right);
 /// let line = Line::from("Hello world!").centered();
@@ -134,13 +134,12 @@ use crate::{
 /// `Line` implements the [`Widget`] trait, which means it can be rendered to a [`Buffer`].
 ///
 /// ```rust
-/// use ratatui::{
+/// use ratatui_core::{
 ///     buffer::Buffer,
 ///     layout::Rect,
 ///     style::{Style, Stylize},
 ///     text::Line,
 ///     widgets::Widget,
-///     Frame,
 /// };
 ///
 /// # fn render(area: Rect, buf: &mut Buffer) {
@@ -148,19 +147,23 @@ use crate::{
 /// let line = Line::from("Hello world!").style(Style::new().yellow().italic());
 /// line.render(area, buf);
 /// # }
+/// ```
 ///
+/// Or you can use the `render_widget` method on the `Frame` in a `Terminal::draw` closure.
+///
+/// ```rust,ignore
+/// # use ratatui::{Frame, layout::Rect, text::Line};
 /// # fn draw(frame: &mut Frame, area: Rect) {
-/// // in a terminal.draw closure
-/// let line = Line::from("Hello world!").style(Style::new().yellow().italic());
+/// let line = Line::from("Hello world!");
 /// frame.render_widget(line, area);
 /// # }
 /// ```
 /// ## Rendering Lines with a Paragraph widget
 ///
-/// Usually apps will use the [`Paragraph`] widget instead of rendering a [`Line`] directly as it
+/// Usually apps will use the `Paragraph` widget instead of rendering a [`Line`] directly as it
 /// provides more functionality.
 ///
-/// ```rust
+/// ```rust,ignore
 /// use ratatui::{
 ///     buffer::Buffer,
 ///     layout::Rect,
@@ -177,7 +180,6 @@ use crate::{
 /// # }
 /// ```
 ///
-/// [`Paragraph`]: crate::widgets::Paragraph
 /// [`Stylize`]: crate::style::Stylize
 #[derive(Default, Clone, Eq, PartialEq, Hash)]
 pub struct Line<'a> {
@@ -241,7 +243,7 @@ impl<'a> Line<'a> {
     /// ```rust
     /// use std::borrow::Cow;
     ///
-    /// use ratatui::text::Line;
+    /// use ratatui_core::text::Line;
     ///
     /// Line::raw("test content");
     /// Line::raw(String::from("test content"));
@@ -272,7 +274,7 @@ impl<'a> Line<'a> {
     /// ```rust
     /// use std::borrow::Cow;
     ///
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     style::{Style, Stylize},
     ///     text::Line,
     /// };
@@ -304,7 +306,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::{style::Stylize, text::Line};
+    /// use ratatui_core::{style::Stylize, text::Line};
     ///
     /// let line = Line::default().spans(vec!["Hello".blue(), " world!".green()]);
     /// let line = Line::default().spans([1, 2, 3].iter().map(|i| format!("Item {}", i)));
@@ -332,7 +334,7 @@ impl<'a> Line<'a> {
     ///
     /// # Examples
     /// ```rust
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     style::{Style, Stylize},
     ///     text::Line,
     /// };
@@ -356,7 +358,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::{layout::Alignment, text::Line};
+    /// use ratatui_core::{layout::Alignment, text::Line};
     ///
     /// let mut line = Line::from("Hi, what's up?");
     /// assert_eq!(None, line.alignment);
@@ -382,7 +384,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::text::Line;
+    /// use ratatui_core::text::Line;
     ///
     /// let line = Line::from("Hi, what's up?").left_aligned();
     /// ```
@@ -400,7 +402,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::text::Line;
+    /// use ratatui_core::text::Line;
     ///
     /// let line = Line::from("Hi, what's up?").centered();
     /// ```
@@ -418,7 +420,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::text::Line;
+    /// use ratatui_core::text::Line;
     ///
     /// let line = Line::from("Hi, what's up?").right_aligned();
     /// ```
@@ -432,7 +434,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::{style::Stylize, text::Line};
+    /// use ratatui_core::{style::Stylize, text::Line};
     ///
     /// let line = Line::from(vec!["Hello".blue(), " world!".green()]);
     /// assert_eq!(12, line.width());
@@ -454,7 +456,7 @@ impl<'a> Line<'a> {
     /// ```rust
     /// use std::iter::Iterator;
     ///
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     style::{Color, Style},
     ///     text::{Line, StyledGrapheme},
     /// };
@@ -498,7 +500,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     style::{Color, Modifier},
     ///     text::Line,
     /// };
@@ -527,7 +529,7 @@ impl<'a> Line<'a> {
     ///
     /// ```rust
     /// # let style = Style::default().yellow();
-    /// use ratatui::{
+    /// use ratatui_core::{
     ///     style::{Style, Stylize},
     ///     text::Line,
     /// };
@@ -559,7 +561,7 @@ impl<'a> Line<'a> {
     /// # Examples
     ///
     /// ```rust
-    /// use ratatui::text::{Line, Span};
+    /// use ratatui_core::text::{Line, Span};
     ///
     /// let mut line = Line::from("Hello, ");
     /// line.push_span(Span::raw("world!"));
