+++ ratatui/Cargo.toml
@@ -15,25 +15,21 @@ edition.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-bitflags = "2.3"
-cassowary = "0.3"
-compact_str = "0.8.0"
+bitflags.workspace = true
 crossterm = { version = "0.28.1", optional = true }
 document-features = { version = "0.2.7", optional = true }
 indoc = "2"
-instability = "0.3.1"
-itertools = "0.13"
-lru = "0.12.0"
-paste = "1.0.2"
+instability.workspace = true
+itertools.workspace = true
 palette = { version = "0.7.6", optional = true }
-serde = { version = "1", optional = true, features = ["derive"] }
-strum = { version = "0.26.3", features = ["derive"] }
+ratatui-core = { workspace = true, features = ["unstable-widget-ref"] }
+serde = { workspace = true, optional = true }
+strum.workspace = true
 termwiz = { version = "0.22.0", optional = true }
 time = { version = "0.3.11", optional = true, features = ["local-offset"] }
-unicode-segmentation = "1.10"
-unicode-truncate = "1"
+unicode-segmentation.workspace = true
 # See <https://github.com/ratatui/ratatui/issues/1271> for information about why we pin unicode-width
-unicode-width = "=0.2.0"
+unicode-width.workspace = true
 
 [target.'cfg(not(windows))'.dependencies]
 # termion is not supported on Windows
@@ -124,13 +120,13 @@ termwiz = ["dep:termwiz"]
 #! The following optional features are available for all backends:
 ## enables serialization and deserialization of style and color types using the [`serde`] crate.
 ## This is useful if you want to save themes to a file.
-serde = ["dep:serde", "bitflags/serde", "compact_str/serde"]
+serde = ["dep:serde", "ratatui-core/serde"]
 
 ## enables the [`border!`] macro.
 macros = []
 
 ## enables conversions from colors in the [`palette`] crate to [`Color`](crate::style::Color).
-palette = ["dep:palette"]
+palette = ["ratatui-core/palette", "dep:palette"]
 
 ## Use terminal scrolling regions to make some operations less prone to
 ## flickering. (i.e. Terminal::insert_before).
@@ -149,7 +145,7 @@ widget-calendar = ["dep:time"]
 ## enables the backend code that sets the underline color.
 ## Underline color is only supported by the [`CrosstermBackend`](backend::CrosstermBackend) backend,
 ## and is not supported on Windows 7.
-underline-color = ["dep:crossterm"]
+underline-color = ["dep:crossterm", "ratatui-core/underline-color"]
 
 #! The following features are unstable and may change in the future:
 
@@ -166,9 +162,12 @@ unstable = [
 ## See [Issue 293](https://github.com/ratatui/ratatui/issues/293) for more details.
 unstable-rendered-line-info = []
 
-## Enables the [`WidgetRef`](widgets::WidgetRef) and [`StatefulWidgetRef`](widgets::StatefulWidgetRef) traits which are experimental and may change in
-## the future.
-unstable-widget-ref = []
+## enables the [`WidgetRef`] and [`StatefulWidgetRef`] traits which are experimental and may change
+## in the future.
+##
+## [`WidgetRef`]: widgets::WidgetRef
+## [`StatefulWidgetRef`]: widgets::StatefulWidgetRef
+unstable-widget-ref = ["ratatui-core/unstable-widget-ref"]
 
 ## Enables getting access to backends' writers.
 unstable-backend-writer = []
